+++ facebook-core/src/main/java/com/facebook/FacebookSdkVersion.java
@@ -21,5 +21,5 @@
 package com.facebook;
 
 final class FacebookSdkVersion {
-  public static final String BUILD = "8.1.0";
+  public static final String BUILD = "8.2.0";
 }